website/requirements.txt
click==6.7
Flask==1.0.2
itsdangerous==0.24
Jinja2==2.10
MarkupSafe==1.0
Werkzeug==0.14.1
Flask-Bootstrap==3.3.7.1
Flask-SocketIO==3.1.0
python-SocketIO==2.1.0
pillow==5.2.0
tensorflow==1.2.1
Jieba==0.38
Gensim==2.0.0
pypinyin==0.23
IPython==7.1.1
requests
matplotlib
